185: move line counting logic up into App component
