+++ src/pages/HomePage/HomePage.tsx
@@ -1,56 +1,31 @@
-import { createSignal, For, Show } from "solid-js";
+import { For, Show } from "solid-js";
 import { A, useNavigate } from "@solidjs/router";
-import { invoke } from "@tauri-apps/api/core";
-import { open } from "@tauri-apps/plugin-dialog";
 import Routes from "../../routes";
 import UploadButton from "../../components/UploadButton/UploadButton";
 import CountingLines from "../../components/CountingLines";
 import { useI18n } from "../../utils/i18n";
 import logAsyncErrors from "../../utils/async/logErrors";
-import {
-  conicGradient,
-  createTree,
-  LINE_TYPE,
-  type Node,
-} from "../../utils/zon";
-import type { Languages } from "../../utils/tokei";
+import { conicGradient, type Node } from "../../utils/zon";
 import Logo from "../../components/Logo";
 import NumberOfLines from "../../components/NumberOfLines";
 import styles from "./HomePage.module.css";
 
-type LandingPageProps = {
+type HomePageProps = {
   reports: Record<string, Node>;
-  setReport: (path: string, root: Node) => void;
+  countLinesInFolder: () => Promise<string | null>;
+  countingPath: string | null;
 };
 
-export default function LandingPage(props: LandingPageProps) {
+export default function HomePage(props: HomePageProps) {
   const navigate = useNavigate();
   const { t } = useI18n();
 
-  const [countingPath, setCountingPath] = createSignal<string | null>();
-
   async function countLinesInFolder() {
-    const path = await open({
-      multiple: false,
-      directory: true,
-    });
+    const path = await props.countLinesInFolder();
 
-    if (path == null) {
-      return;
+    if (path != null) {
+      navigate(Routes.Report.getLocation(path));
     }
-
-    setCountingPath(path);
-
-    const languages = await invoke("count_lines", { path: path });
-    const reportRoot = createTree(path, languages as Languages, [
-      LINE_TYPE.BLANKS,
-      LINE_TYPE.CODE,
-      LINE_TYPE.COMMENTS,
-    ]);
-
-    props.setReport(path, reportRoot);
-    setCountingPath(null);
-    navigate(Routes.Report.getLocation(path));
   }
 
   return (
@@ -105,7 +80,7 @@ export default function LandingPage(props: LandingPageProps) {
           <UploadButton
             countLinesInFolder={logAsyncErrors(countLinesInFolder)}
           />
-          <Show when={countingPath()}>
+          <Show when={props.countingPath}>
             {(definedPath) => <CountingLines path={definedPath()} />}
           </Show>
         </div>
